+++ src/lib.rs
@@ -26,6 +26,8 @@ pub mod rfc6186;
 pub mod rfc6764;
 #[cfg(feature = "rfc8620")]
 pub mod rfc8620;
+#[cfg(any(feature = "rfc8620", feature = "search"))]
+pub mod rfc9110;
 #[cfg(feature = "search")]
 pub mod search;
 #[cfg(any(